Transformation: refactor publisher files into own sub-directory
