+++ transformation/src/pipeline-config/publisher/amqpExecutionResultPublisher.ts
